+++ test/support/mocks.ex
@@ -1,3 +1,4 @@
 Mox.defmock(Rheo.Backend.Mongo.ClientMock, for: Rheo.Backend.Mongo.Client)
 Mox.defmock(Rheo.Backend.Mock, for: Rheo.Backend)
 Mox.defmock(Rheo.Backend.Redis.ClientMock, for: Rheo.Backend.Redis.Client)
+Mox.defmock(Rheo.Backend.Mnesia.StoreMock, for: Rheo.Backend.Mnesia.Store)